+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - utils.py: added convert_rhntw_to_rhcbci, the inverse of convert_rhcbci_to_rhntw, and a common function between the two: compute_R_rhntw_rhcbci to compute the rotation matrix they both use
 - private_src/: LBG1_LG6 environments and lbg1_private_utils
 - lbg1_base.py: adding env state vars for lady and bandit capture to enable score modifiers upon capture
+- run_tests.py and pyproject.toml: script for running aiaa competition tests
 
 ### Changed: